routes

+++ src/configuration/rest_err/rest.err.go
@@ -3,14 +3,14 @@ package rest_err
 import "net/http"
 
 type RestErr struct {
-	Message string  `json:"message"`
-	Err string `json:"error"`
-	Code int64 `json:"code"`
-	Causes []Causes `json:"causes"`
+	Message string   `json:"message"`
+	Err     string   `json:"error"`
+	Code    int    `json:"code"`
+	Causes  []Causes `json:"causes"`
 }
 
 type Causes struct {
-	Field string `json:"field"`
+	Field   string `json:"field"`
 	Message string `json:"message"`
 }
 
@@ -19,54 +19,52 @@ func (r *RestErr) Error() string {
 	return r.Message
 }
 
-
-func NewRestErr(message, err string, code int64, causes [] Causes) *RestErr{
+func NewRestErr(message, err string, code int, causes []Causes) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: err,
-		Code: code,
-		Causes: causes,
-
+		Err:     err,
+		Code:    code,
+		Causes:  causes,
 	}
 }
 
 func NewBadRequestError(message string) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: "bad_request",
-		Code: http.StatusBadRequest,
+		Err:     "bad_request",
+		Code:    http.StatusBadRequest,
 	}
 }
 
 func NewBadRequestValidationError(message string, causes []Causes) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: "bad_request",
-		Code: http.StatusBadRequest,
-		Causes: causes,
+		Err:     "bad_request",
+		Code:    http.StatusBadRequest,
+		Causes:  causes,
 	}
 }
 
-func NewInternalServerError(message string) *RestErr{
+func NewInternalServerError(message string) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: "internal_server_error",
-		Code: http.StatusInternalServerError,
+		Err:     "internal_server_error",
+		Code:    http.StatusInternalServerError,
 	}
 }
 
-func NewNotFoundError(message string) *RestErr{
+func NewNotFoundError(message string) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: "not_found",
-		Code: http.StatusInternalServerError,
+		Err:     "not_found",
+		Code:    http.StatusInternalServerError,
 	}
 }
 
-func NewForbidden(message string) *RestErr{
+func NewForbidden(message string) *RestErr {
 	return &RestErr{
 		Message: message,
-		Err: "forbidden",
-		Code: http.StatusForbidden,
+		Err:     "forbidden",
+		Code:    http.StatusForbidden,
 	}
 }
